Crate wasmtime_wasi_nn

Modules

backend

Define the Rust interface a backend must implement in order to be used by this crate. The Box<dyn ...> types returned by these interfaces allow implementations to maintain backend-specific state between calls.

testing

This is testing-specific code–it is public only so that it can be accessible both in unit and integration tests.

wit

Implements the wasi-nn API for the WIT (“preview2”) ABI.

witx

Implements the wasi-nn API for the WITX (“preview1”) ABI.

Macros

check_test

Early-return from a test if the test environment is not met. If the CI or FORCE_WASINN_TEST_CHECK environment variables are set, though, this will return an error instead.

Structs

Backend

A machine learning backend.

ExecutionContext

A backend-defined execution context.

Graph

A backend-defined graph (i.e., ML model).

InMemoryRegistry

Registry

A container for graphs.

WasiNnCtx

Capture the state necessary for calling into the backend ML libraries.

Traits

GraphRegistry

Functions

preload

Construct an in-memory registry from the available backends and a list of (<backend name>, <graph directory>). This assumes graphs can be loaded from a local directory, which is a safe assumption currently for the current model types.