Module wasmtime_wasi_nn::wit

Implements the wasi-nn API for the WIT (“preview2”) ABI.

Note that wasi-nn is not yet included in an official “preview2” world (though it could be) so by “preview2” here we mean that this can be called with the component model’s canonical ABI.

This module exports its types for use throughout the crate and the ML object, which exposes ML::add_to_linker. To implement all of this, this module proceeds in steps:

generate all of the WIT glue code into a gen::* namespace

wire up the gen::* glue to the context state, delegating actual computation to a Backend

convert some types

Modules

types

Structs

GraphEncodingParseError

ML